+++ ios/Plugin/AppsFlyerPlugin.m
@@ -45,6 +45,7 @@ CAP_PLUGIN(AppsFlyerPlugin, "AppsFlyerPlugin",
            CAP_PLUGIN_METHOD(setConsentDataV2, CAPPluginReturnNone);
            CAP_PLUGIN_METHOD(isSDKStarted, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(isSDKStopped, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(validateAndLogInAppPurchaseV2, CAPPluginReturnPromise);